Make v2 versions of all the functions
+ Rename affected files and test files to *_v2
